requirements.txt
langchain==0.3.8
langchain-community==0.3.8
langchain-core==0.3.21
langchain-openai==0.2.10
langchain-text-splitters==0.3.2
langgraph
chromadb==0.5.20
python-dotenv==1.0.1
streamlit==1.40.0
mysql-connector-python==9.2.0